posix: deprecate PTHREAD_IPC for POSIX_THREADS_BASE
Zephyr's POSIX API is moving toward using the standard
nomenclature from IEEE 1003.1-2017 for as much as possible. In
particular, we want to have consistent naming between Zephyr's
POSIX API Kconfig options and the naming for POSIX Options and
Option Groups.

The Kconfig option CONFIG_PTHREAD_IPC has been used for a very
long time for a variety of different purposes. However, the
standard Option Group for POSIX Threads is POSIX_THREADS_BASE,
which would translate to the Kconfig option
CONFIG_POSIX_THREADS_BASE.

The POSIX_THREADS_BASE Option Group already includes mutexes,
condition variables, and thread-specific storage (keys). So with
this change, we also deprecate the redundant Kconfig variables
that do not have a corresponding match in the standard.

Additionally, create Kconfig variables for those configurables
which we are lacking:

- CONFIG_POSIX_SEMAPHORES
- CONFIG_POSIX_THREAD_PRIORITY_SCHEDULING
- CONFIG_POSIX_THREADS_EXT
- CONFIG_POSIX_THREAD_ATTR_STACKSIZE
- CONFIG_POSIX_THREAD_ATTR_STACKADDR

Some Kconfig variables were renamed to more properly match the
spec:

- CONFIG_PTHREAD_KEY_MAX_COUNT -> CONFIG_POSIX_THREAD_KEYS_MAX
- CONFIG_PTHREAD_MAX_COUNT -> CONFIG_POSIX_THREAD_THREADS_MAX

Note: POSIX also requires the Option _POSIX_THREADS as part of
the mandatory System Interfaces, so when
CONFIG_POSIX_THREADS_BASE=y, the implication is that
_POSIX_THREADS is defined to be 200809L, or whatever the
appropriate conformance level demands.
