+++ include/zephyr/posix/pthread.h
@@ -417,9 +417,7 @@ int pthread_attr_getscope(const pthread_attr_t *attr, int *contentionscope);
 int pthread_attr_setscope(pthread_attr_t *attr, int contentionscope);
 int pthread_attr_getinheritsched(const pthread_attr_t *attr, int *inheritsched);
 int pthread_attr_setinheritsched(pthread_attr_t *attr, int inheritsched);
-#ifdef CONFIG_PTHREAD_IPC
 int pthread_once(pthread_once_t *once, void (*initFunc)(void));
-#endif
 FUNC_NORETURN void pthread_exit(void *retval);
 int pthread_join(pthread_t thread, void **status);
 int pthread_cancel(pthread_t pthread);
@@ -502,8 +500,6 @@ int pthread_setname_np(pthread_t thread, const char *name);
  */
 int pthread_getname_np(pthread_t thread, char *name, size_t len);
 
-#ifdef CONFIG_PTHREAD_IPC
-
 /**
  * @brief Destroy a pthread_spinlock_t.
  *
@@ -538,8 +534,6 @@ int pthread_spin_trylock(pthread_spinlock_t *lock);
  * See IEEE 1003.1
  */
 int pthread_spin_unlock(pthread_spinlock_t *lock);
-
-#endif
 
 #ifdef __cplusplus
 }
