+++ include/zephyr/posix/unistd.h
@@ -38,7 +38,7 @@ extern "C" {
 /* Constants for Options and Option Groups */
 #define _POSIX_ADVISORY_INFO		  (-1L)
 #define _POSIX_ASYNCHRONOUS_IO Z_SC_VAL_IFDEF(CONFIG_POSIX_ASYNCHRONOUS_IO, _POSIX_VERSION)
-#define _POSIX_BARRIERS			  Z_SC_VAL_IFDEF(CONFIG_PTHREAD_BARRIER, _POSIX_VERSION)
+#define _POSIX_BARRIERS			  Z_SC_VAL_IFDEF(CONFIG_POSIX_BARRIERS, _POSIX_VERSION)
 #define _POSIX_CHOWN_RESTRICTED		  (0)
 #define _POSIX_CLOCK_SELECTION		  Z_SC_VAL_IFDEF(CONFIG_POSIX_CLOCK, _POSIX_VERSION)
 #define _POSIX_CPUTIME			  (-1L)
@@ -49,39 +49,39 @@ extern "C" {
 #define _POSIX_MEMLOCK			  (-1L)
 #define _POSIX_MEMLOCK_RANGE		  (-1L)
 #define _POSIX_MEMORY_PROTECTION	  (-1L)
-#define _POSIX_MESSAGE_PASSING		  Z_SC_VAL_IFDEF(CONFIG_POSIX_MQUEUE, _POSIX_VERSION)
+#define _POSIX_MESSAGE_PASSING		  Z_SC_VAL_IFDEF(CONFIG_POSIX_MESSAGE_PASSING, _POSIX_VERSION)
 #define _POSIX_MONOTONIC_CLOCK		  Z_SC_VAL_IFDEF(CONFIG_POSIX_CLOCK, _POSIX_VERSION)
 #define _POSIX_NO_TRUNC			  (0)
 #define _POSIX_PRIORITIZED_IO		  (-1L)
 #define _POSIX_PRIORITY_SCHEDULING Z_SC_VAL_IFDEF(CONFIG_POSIX_PRIORITY_SCHEDULING, _POSIX_VERSION)
 #define _POSIX_RAW_SOCKETS		  Z_SC_VAL_IFDEF(CONFIG_NET_SOCKETS_PACKET, _POSIX_VERSION)
-#define _POSIX_READER_WRITER_LOCKS	  Z_SC_VAL_IFDEF(CONFIG_PTHREAD_IPC, _POSIX_VERSION)
+#define _POSIX_READER_WRITER_LOCKS Z_SC_VAL_IFDEF(CONFIG_POSIX_READER_WRITER_LOCKS, _POSIX_VERSION)
 #define _POSIX_REALTIME_SIGNALS		  (-1L)
 #define _POSIX_REGEXP			  (-1L)
 #define _POSIX_SAVED_IDS		  (-1L)
-#define _POSIX_SEMAPHORES		  Z_SC_VAL_IFDEF(CONFIG_PTHREAD_IPC, _POSIX_VERSION)
+#define _POSIX_SEMAPHORES		  Z_SC_VAL_IFDEF(CONFIG_POSIX_SEMAPHORES, _POSIX_VERSION)
 #define _POSIX_SHARED_MEMORY_OBJECTS	  (-1L)
 #define _POSIX_SHELL			  (-1L)
 #define _POSIX_SPAWN			  (-1L)
-#define _POSIX_SPIN_LOCKS		  Z_SC_VAL_IFDEF(CONFIG_PTHREAD_SPINLOCK, _POSIX_VERSION)
+#define _POSIX_SPIN_LOCKS		  Z_SC_VAL_IFDEF(CONFIG_POSIX_SPIN_LOCKS, _POSIX_VERSION)
 #define _POSIX_SPORADIC_SERVER		  (-1L)
 #define _POSIX_SYNCHRONIZED_IO		  (-1L)
-#define _POSIX_THREAD_ATTR_STACKADDR	  Z_SC_VAL_IFDEF(CONFIG_PTHREAD_IPC, _POSIX_VERSION)
-#define _POSIX_THREAD_ATTR_STACKSIZE	  Z_SC_VAL_IFDEF(CONFIG_PTHREAD_IPC, _POSIX_VERSION)
+#define _POSIX_THREAD_ATTR_STACKADDR                                                               \
+	Z_SC_VAL_IFDEF(CONFIG_POSIX_THREAD_ATTR_STACKADDR, _POSIX_VERSION)
+#define _POSIX_THREAD_ATTR_STACKSIZE                                                               \
+	Z_SC_VAL_IFDEF(CONFIG_POSIX_THREAD_ATTR_STACKSIZE, _POSIX_VERSION)
 #define _POSIX_THREAD_CPUTIME		  (-1L)
 #define _POSIX_THREAD_PRIO_INHERIT	  _POSIX_VERSION
 #define _POSIX_THREAD_PRIO_PROTECT	  (-1L)
-#define _POSIX_THREAD_PRIORITY_SCHEDULING Z_SC_VAL_IFDEF(CONFIG_PTHREAD_IPC, _POSIX_VERSION)
+#define _POSIX_THREAD_PRIORITY_SCHEDULING                                                          \
+	Z_SC_VAL_IFDEF(CONFIG_POSIX_THREAD_PRIORITY_SCHEDULING, _POSIX_VERSION)
 #define _POSIX_THREAD_PROCESS_SHARED	  (-1L)
 #define _POSIX_THREAD_ROBUST_PRIO_INHERIT (-1L)
 #define _POSIX_THREAD_ROBUST_PRIO_PROTECT (-1L)
-#define _POSIX_THREAD_SAFE_FUNCTIONS	  Z_SC_VAL_IFDEF(CONFIG_PTHREAD_IPC, _POSIX_VERSION)
+#define _POSIX_THREAD_SAFE_FUNCTIONS                                                               \
+	Z_SC_VAL_IFDEF(CONFIG_POSIX_THREAD_SAFE_FUNCTIONS, _POSIX_VERSION)
 #define _POSIX_THREAD_SPORADIC_SERVER	  (-1L)
-
-#ifndef _POSIX_THREADS
-#define _POSIX_THREADS			  Z_SC_VAL_IFDEF(CONFIG_PTHREAD_IPC, _POSIX_VERSION)
-#endif
-
+#define _POSIX_THREADS			  Z_SC_VAL_IFDEF(CONFIG_POSIX_THREADS_BASE, _POSIX_VERSION)
 #define _POSIX_TIMEOUTS			  Z_SC_VAL_IFDEF(CONFIG_POSIX_CLOCK, _POSIX_VERSION)
 #define _POSIX_TIMERS			  Z_SC_VAL_IFDEF(CONFIG_POSIX_CLOCK, _POSIX_VERSION)
 #define _POSIX_TRACE			  (-1L)
@@ -154,8 +154,9 @@ extern "C" {
 #define _POSIX_SYMLINK_MAX		    (255)
 #define _POSIX_SYMLOOP_MAX		    (8)
 #define _POSIX_THREAD_DESTRUCTOR_ITERATIONS (4)
-#define _POSIX_THREAD_KEYS_MAX		    (128)
-#define _POSIX_THREAD_THREADS_MAX	    (64)
+#define _POSIX_THREAD_THREADS_MAX	COND_CODE_1(CONFIG_POSIX_THREADS_BASE,	      \
+						  (CONFIG_POSIX_THREAD_THREADS_MAX), \
+						  (0))
 #define _POSIX_TIMER_MAX		    (32)
 #define _POSIX_TRACE_EVENT_NAME_MAX	    (30)
 #define _POSIX_TRACE_NAME_MAX		    (8)
@@ -207,12 +208,11 @@ extern "C" {
 #endif
 
 #define PTHREAD_DESTRUCTOR_ITERATIONS _POSIX_THREAD_DESTRUCTOR_ITERATIONS
-#define PTHREAD_KEYS_MAX	      COND_CODE_1(CONFIG_PTHREAD_IPC,		  \
-						  (CONFIG_MAX_PTHREAD_KEY_COUNT), \
-						  (_POSIX_THREAD_KEYS_MAX))
-#define PTHREAD_THREADS_MAX	      COND_CODE_1(CONFIG_PTHREAD_IPC,	      \
-						  (CONFIG_MAX_PTHREAD_COUNT), \
+#define _POSIX_THREAD_KEYS_MAX	      COND_CODE_1(CONFIG_POSIX_THREADS_BASE,		  \
+						  (CONFIG_POSIX_THREAD_KEYS_MAX), \
 						  (0))
+#define PTHREAD_KEYS_MAX          _POSIX_THREAD_KEYS_MAX
+#define PTHREAD_THREADS_MAX	      _PTHREAD_THREADS_MAX
 #define SEM_NSEMS_MAX		      _POSIX_SEM_NSEMS_MAX
 #define SEM_VALUE_MAX		      CONFIG_SEM_VALUE_MAX
 #define SIGQUEUE_MAX		      _POSIX_SIGQUEUE_MAX
