+++ lib/posix/options/CMakeLists.txt
@@ -24,14 +24,6 @@ if(CONFIG_POSIX_SIGNAL)
   )
 endif()
 
-if(CONFIG_POSIX_API OR CONFIG_PTHREAD_IPC OR CONFIG_POSIX_CLOCK OR
-  CONFIG_POSIX_MQUEUE OR CONFIG_POSIX_FS OR CONFIG_EVENTFD OR CONFIG_GETOPT)
-  # This is a temporary workaround so that Newlib declares the appropriate
-  # types for us. POSIX features to be formalized as part of #51211
-  zephyr_compile_options($<$<COMPILE_LANGUAGE:C>:-D_POSIX_THREADS>)
-  zephyr_compile_options($<$<COMPILE_LANGUAGE:CXX>:-D_POSIX_THREADS>)
-endif()
-
 zephyr_library()
 add_subdirectory_ifdef(CONFIG_GETOPT getopt)
 zephyr_library_sources_ifdef(CONFIG_EVENTFD eventfd.c)
@@ -45,23 +37,28 @@ zephyr_library_sources_ifdef(CONFIG_POSIX_CLOCK sleep.c)
 zephyr_library_sources_ifdef(CONFIG_POSIX_CONFSTR confstr.c)
 zephyr_library_sources_ifdef(CONFIG_POSIX_ENV env.c)
 zephyr_library_sources_ifdef(CONFIG_POSIX_FS fs.c)
-zephyr_library_sources_ifdef(CONFIG_POSIX_MQUEUE mqueue.c)
+zephyr_library_sources_ifdef(CONFIG_POSIX_MESSAGE_PASSING mqueue.c)
 zephyr_library_sources_ifdef(CONFIG_POSIX_NETWORKING net.c)
 zephyr_library_sources_ifdef(CONFIG_POSIX_PUTMSG stropts.c)
 zephyr_library_sources_ifdef(CONFIG_POSIX_SIGNAL signal.c ${STRSIGNAL_TABLE_H})
 zephyr_library_sources_ifdef(CONFIG_POSIX_SYSCONF_IMPL_FULL sysconf.c)
 zephyr_library_sources_ifdef(CONFIG_POSIX_SYSLOG syslog.c)
+zephyr_library_sources_ifdef(CONFIG_POSIX_THREADS_BASE
+  cond.c
+  key.c
+  mutex.c
+  pthread.c
+)
+# zephyr_compile_options_ifdef(CONFIG_POSIX_THREADS_BASE
+#   $<$<COMPILE_LANGUAGE:C>:-U_POSIX_THREADS -D_POSIX_THREADS=200809L>
+#   $<$<COMPILE_LANGUAGE:CXX>:-U_POSIX_THREADS -D_POSIX_THREADS=200809L>
+# )
 zephyr_library_sources_ifdef(CONFIG_POSIX_UNAME uname.c)
-zephyr_library_sources_ifdef(CONFIG_PTHREAD_IPC _common.c)
-zephyr_library_sources_ifdef(CONFIG_PTHREAD_BARRIER barrier.c)
-zephyr_library_sources_ifdef(CONFIG_PTHREAD_COND cond.c)
-zephyr_library_sources_ifdef(CONFIG_PTHREAD_KEY key.c)
-zephyr_library_sources_ifdef(CONFIG_PTHREAD_MUTEX mutex.c)
-zephyr_library_sources_ifdef(CONFIG_PTHREAD pthread.c)
-zephyr_library_sources_ifdef(CONFIG_PTHREAD_RWLOCK rwlock.c)
+zephyr_library_sources_ifdef(CONFIG_POSIX_BARRIERS barrier.c)
+zephyr_library_sources_ifdef(CONFIG_POSIX_READER_WRITER_LOCKS rwlock.c)
 zephyr_library_sources_ifdef(CONFIG_POSIX_PRIORITY_SCHEDULING sched.c)
-zephyr_library_sources_ifdef(CONFIG_PTHREAD_IPC semaphore.c)
-zephyr_library_sources_ifdef(CONFIG_PTHREAD_SPINLOCK spinlock.c)
+zephyr_library_sources_ifdef(CONFIG_POSIX_SEMAPHORES semaphore.c)
+zephyr_library_sources_ifdef(CONFIG_POSIX_SPIN_LOCKS spinlock.c)
 zephyr_library_sources_ifdef(CONFIG_TIMER timer.c)
 
 zephyr_library_include_directories(
