+++ lib/posix/options/Kconfig
@@ -13,22 +13,9 @@ config POSIX_API
 	  Enable mostly-standards-compliant implementations of
 	  various POSIX (IEEE 1003.1) APIs.
 
-if POSIX_CLOCK
-
-config PTHREAD_IPC
-	bool "POSIX pthread IPC API"
-	default y if POSIX_API
-	help
-	  This enables a mostly-standards-compliant implementation of
-	  the pthread mutex, condition variable and barrier IPC
-	  mechanisms.
-
-endif # POSIX_CLOCK
-
 rsource "Kconfig.aio"
 rsource "Kconfig.barrier"
 rsource "Kconfig.clock"
-rsource "Kconfig.cond"
 rsource "Kconfig.confstr"
 rsource "Kconfig.env"
 rsource "Kconfig.eventfd"
@@ -37,9 +24,7 @@ rsource "Kconfig.fnmatch"
 rsource "Kconfig.fs"
 rsource "Kconfig.getentropy"
 rsource "Kconfig.getopt"
-rsource "Kconfig.key"
 rsource "Kconfig.mqueue"
-rsource "Kconfig.mutex"
 rsource "Kconfig.net"
 rsource "Kconfig.pthread"
 rsource "Kconfig.rwlock"
@@ -52,5 +37,9 @@ rsource "Kconfig.sysconf"
 rsource "Kconfig.syslog"
 rsource "Kconfig.timer"
 rsource "Kconfig.uname"
+rsource "Kconfig.xsi"
+
+# To be removed at the end of the deprecation process
+rsource "Kconfig.deprecated"
 
 endmenu # "POSIX Options"
