--- lib/posix/options/Kconfig.cond
@@ -1,9 +1,0 @@
-# Copyright (c) 2017 Intel Corporation
-# Copyright (c) 2023 Meta
-#
-# SPDX-License-Identifier: Apache-2.0
-
-TYPE = PTHREAD_COND
-type = pthread_cond_t
-type-function = pthread_cond_wait
-rsource "Kconfig.template.pooled_ipc_type"
