+++ lib/posix/options/Kconfig.env
@@ -9,6 +9,10 @@ config POSIX_ENV
 	help
 	  Select this option to add support for environment variables.
 
+if POSIX_ENV
+
 module = POSIX_ENV
 module-str = POSIX env logging
 source "subsys/logging/Kconfig.template.log_config"
+
+endif # POSIX_ENV
