--- lib/posix/options/Kconfig.key
@@ -1,9 +1,0 @@
-# Copyright (c) 2018 Intel Corporation
-# Copyright (c) 2023 Meta
-#
-# SPDX-License-Identifier: Apache-2.0
-
-TYPE = PTHREAD_KEY
-type = pthread_key_t
-type-function = pthread_setspecific
-rsource "Kconfig.template.pooled_ipc_type"
