--- lib/posix/options/Kconfig.mutex
@@ -1,9 +1,0 @@
-# Copyright (c) 2017 Intel Corporation
-# Copyright (c) 2023 Meta
-#
-# SPDX-License-Identifier: Apache-2.0
-
-TYPE = PTHREAD_MUTEX
-type = pthread_mutex_t
-type-function = pthread_mutex_lock
-rsource "Kconfig.template.pooled_ipc_type"
