+++ lib/posix/options/Kconfig.pthread
@@ -3,12 +3,51 @@
 #
 # SPDX-License-Identifier: Apache-2.0
 
-TYPE = PTHREAD
-type = pthread_t
-type-function = pthread_create
-rsource "Kconfig.template.pooled_ipc_type"
+menuconfig POSIX_THREADS_BASE
+	bool "POSIX thread support"
+	default y if POSIX_API
+	help
+	  Select 'y' here to enable POSIX threads, mutexes, condition variables, and thread-specific
+	  storage.
 
-if PTHREAD
+	  For more information please see
+	  https://pubs.opengroup.org/onlinepubs/9699919799/xrat/V4_subprofiles.html
+
+if POSIX_THREADS_BASE
+
+config POSIX_THREAD_THREADS_MAX
+	int "Maximum number of pthread_t"
+	default 5
+	help
+	  Maximum simultaneously active pthread_t in a POSIX application.
+
+	  For more information, see
+	  https://pubs.opengroup.org/onlinepubs/9699919799/basedefs/limits.h.html
+
+config MAX_PTHREAD_MUTEX_COUNT
+	int "Maximum number of pthread_mutex_t"
+	default 5
+	help
+	  Maximum simultaneously active pthread_mutex_t in a POSIX application.
+
+	  Note: this is a non-standard option.
+
+config MAX_PTHREAD_COND_COUNT
+	int "Maximum number of pthread_cond_t"
+	default 5
+	help
+	  Maximum simultaneously active pthread_cond_t in a POSIX application.
+
+	  Note: this is a non-standard option.
+
+config POSIX_THREAD_KEYS_MAX
+	int "Maximum number of pthread_key_t"
+	default 5
+	help
+	  Maximum simultaneously active pthread_key_t in a POSIX application.
+
+	  For more information, see
+	  https://pubs.opengroup.org/onlinepubs/9699919799/basedefs/limits.h.html
 
 config PTHREAD_RECYCLER_DELAY_MS
 	int "Delay for reclaiming dynamic pthread stacks (ms)"
@@ -26,6 +65,53 @@ config PTHREAD_RECYCLER_DELAY_MS
 
 	  Note: this option should be considered temporary and will likely be
 	  removed once a more synchronous solution is available.
+
+config POSIX_THREAD_ATTR_STACKADDR
+	bool "Support getting and setting POSIX thread stack addresses"
+	help
+	  Enable this option to use pthread_attr_getstackaddr() and
+	  pthread_attr_setstackaddr().
+
+	  This option was removed in IEEE 1003.1-2017 in favour of
+	  pthread_attr_getstack() and pthread_attr_setstack().
+
+	  For more information, please see
+	  https://pubs.opengroup.org/onlinepubs/009696799/functions/pthread_attr_getstackaddr.html
+	  https://pubs.opengroup.org/onlinepubs/9699919799/xrat/V4_xsh_chap03.html
+	  https://pubs.opengroup.org/onlinepubs/9699919799/functions/V2_chap02.html#tag_15_09_08
+
+config POSIX_THREAD_ATTR_STACKSIZE
+	bool "Support getting and setting POSIX thread stack sizes"
+	default y
+	help
+	  Enable this option to use pthread_attr_getstacksize() or
+	  pthread_attr_setstacksize().
+
+	  For more information, please see
+	  https://pubs.opengroup.org/onlinepubs/009696699/functions/pthread_attr_getstacksize.html
+
+config POSIX_THREADS_EXT
+	bool "Extended POSIX thread support"
+	default y
+	help
+	  Enable this option to use pthread_attr_getguardsize(), pthread_attr_setguardsize(),
+	  pthread_mutexattr_gettype(), or pthread_mutexattr_settype().
+
+	  For more information, please see
+	  https://pubs.opengroup.org/onlinepubs/9699919799/xrat/V4_subprofiles.html
+
+config POSIX_THREAD_PRIORITY_SCHEDULING
+	bool "Run POSIX threads with different priorities and schedulers"
+	default y
+	help
+	  Enabling this option allows the application to configure different priorities and
+	  scheduling algorithms for different threads via functions such as pthread_setschedparam()
+	  and pthread_setschedprio(). This is required for Realtime Threads and Advanced Realtime
+	  Threads.
+
+	  For more information, please see
+	  https://pubs.opengroup.org/onlinepubs/9699919799/basedefs/V1_chap02.html#tag_02_01_06
+	  https://man7.org/linux/man-pages/man7/posixoptions.7.html
 
 config POSIX_PTHREAD_ATTR_STACKSIZE_BITS
 	int "Significant bits for pthread_attr_t stacksize"
@@ -59,4 +145,20 @@ config POSIX_PTHREAD_ATTR_GUARDSIZE_DEFAULT
 	  facilitate a more dynamic approach to guard areas (via software or
 	  hardware) but for now it simply increases the size of thread stacks.
 
-endif
+module = PTHREAD
+module-str = POSIX thread
+source "subsys/logging/Kconfig.template.log_config"
+
+module = PTHREAD_MUTEX
+module-str = POSIX mutex
+source "subsys/logging/Kconfig.template.log_config"
+
+module = PTHREAD_COND
+module-str = POSIX condition variable
+source "subsys/logging/Kconfig.template.log_config"
+
+module = PTHREAD_KEY
+module-str = POSIX thread-specific data
+source "subsys/logging/Kconfig.template.log_config"
+
+endif # POSIX_THREADS_BASE
