+++ lib/posix/options/Kconfig.rwlock
@@ -2,7 +2,27 @@
 #
 # SPDX-License-Identifier: Apache-2.0
 
-TYPE = PTHREAD_RWLOCK
-type = pthread_rwlock_t
-type-function = pthread_rwlock_timedrdlock
-rsource "Kconfig.template.pooled_ipc_type"
+menuconfig POSIX_READER_WRITER_LOCKS
+	bool "POSIX reader-writer locks"
+	default y if POSIX_API
+	help
+	  Select 'y' here to enable POSIX reader-writer locks.
+
+	  For more information please see
+	  https://pubs.opengroup.org/onlinepubs/9699919799/xrat/V4_subprofiles.html
+
+if POSIX_READER_WRITER_LOCKS
+
+config MAX_PTHREAD_RWLOCK_COUNT
+	int "Maximum number of pthread_rwlock_t"
+	default 5
+	help
+	  Maximum simultaneously active pthread_rwlock_t in a POSIX application.
+
+	  Note: this is a non-standard option.
+
+module = PTHREAD_RWLOCK
+module-str = POSIX Reader-Writer Locks
+source "subsys/logging/Kconfig.template.log_config"
+
+endif # POSIX_READER_WRITER_LOCKS
