+++ lib/posix/options/Kconfig.sched
@@ -4,8 +4,6 @@
 
 config POSIX_PRIORITY_SCHEDULING
 	bool "Priority scheduling"
-	default y if PTHREAD
 	default y if POSIX_API
-	depends on PTHREAD
 	help
 	  This enables POSIX scheduling APIs (_POSIX_PRIORITY_SCHEDULING).
