+++ lib/posix/options/Kconfig.spinlock
@@ -2,7 +2,27 @@
 #
 # SPDX-License-Identifier: Apache-2.0
 
-TYPE = PTHREAD_SPINLOCK
-type = pthread_spinlock_t
-type-function = pthread_spin_lock
-rsource "Kconfig.template.pooled_ipc_type"
+menuconfig POSIX_SPIN_LOCKS
+	bool "POSIX spin locks"
+	default y if POSIX_API
+	help
+	  Select 'y' here to enable POSIX spin locks.
+
+	  For more information please see
+	  https://pubs.opengroup.org/onlinepubs/9699919799/xrat/V4_subprofiles.html
+
+if POSIX_SPIN_LOCKS
+
+config MAX_PTHREAD_SPINLOCK_COUNT
+	int "Maximum number of pthread_spinlock_t"
+	default 5
+	help
+	  Maximum simultaneously active pthread_spinlock_t in a POSIX application.
+
+	  Note: this is a non-standard option.
+
+module = PTHREAD_SPINLOCK
+module-str = POSIX Spin Locks
+source "subsys/logging/Kconfig.template.log_config"
+
+endif # POSIX_SPIN_LOCKS
