+++ lib/posix/options/Kconfig.xsi
@@ -1,0 +1,14 @@
+# Copyright (c) 2024, Tenstorrent AI ULC
+#
+# SPDX-License-Identifier: Apache-2.0
+
+config XSI_THREADS_EXT
+	bool "X/Open Systems Interface threads extensions"
+	default y if POSIX_API
+	depends on POSIX_THREADS_BASE
+	help
+	  Enable this option to use pthread_attr_getstack(), pthread_attr_setstack(),
+	  pthread_getconcurrency(), or pthread_setconcurrency().
+
+	  For more information, please see
+	  https://pubs.opengroup.org/onlinepubs/9699919799/xrat/V4_subprofiles.html
