+++ lib/posix/options/posix_internal.h
@@ -13,6 +13,7 @@
 #include <zephyr/kernel.h>
 #include <zephyr/posix/pthread.h>
 #include <zephyr/posix/signal.h>
+#include <zephyr/posix/time.h>
 #include <zephyr/sys/dlist.h>
 #include <zephyr/sys/slist.h>
 
@@ -97,6 +98,28 @@ static inline uint32_t mark_pthread_obj_initialized(uint32_t obj)
 static inline uint32_t mark_pthread_obj_uninitialized(uint32_t obj)
 {
 	return obj & ~PTHREAD_OBJ_MASK_INIT;
+}
+
+static inline int64_t timespec_to_timeoutms(const struct timespec *abstime)
+{
+	int64_t milli_secs, secs, nsecs;
+	struct timespec curtime;
+
+	/* FIXME: Zephyr does have CLOCK_REALTIME to get time.
+	 * As per POSIX standard time should be calculated wrt CLOCK_REALTIME.
+	 * Zephyr deviates from POSIX 1003.1 standard on this aspect.
+	 */
+	clock_gettime(CLOCK_MONOTONIC, &curtime);
+	secs = abstime->tv_sec - curtime.tv_sec;
+	nsecs = abstime->tv_nsec - curtime.tv_nsec;
+
+	if (secs < 0 || (secs == 0 && nsecs < NSEC_PER_MSEC)) {
+		milli_secs = 0;
+	} else {
+		milli_secs =  secs * MSEC_PER_SEC + nsecs / NSEC_PER_MSEC;
+	}
+
+	return milli_secs;
 }
 
 struct posix_thread *to_posix_thread(pthread_t pth);
