+++ tests/posix/common/prj.conf
@@ -1,10 +1,6 @@
-CONFIG_PTHREAD_IPC=y
 CONFIG_POSIX_API=y
-CONFIG_MAX_PTHREAD_COUNT=6
+CONFIG_POSIX_THREAD_THREADS_MAX=6
 CONFIG_ZTEST=y
-CONFIG_SEM_VALUE_MAX=32767
-CONFIG_POSIX_MQUEUE=y
-CONFIG_POSIX_PRIORITY_SCHEDULING=y
 CONFIG_HEAP_MEM_POOL_SIZE=4096
 CONFIG_THREAD_NAME=y
 CONFIG_DYNAMIC_THREAD=y
