+++ tests/posix/common/src/sysconf.c
@@ -22,7 +22,7 @@ ZTEST(posix_apis, test_posix_sysconf)
 
 	/* SC that value depends on target's configuration */
 	ret = sysconf(_SC_SEMAPHORES);
-	if (IS_ENABLED(CONFIG_PTHREAD_IPC)) {
+	if (IS_ENABLED(CONFIG_POSIX_SEMAPHORES)) {
 		zassert_equal(ret,
 			      _POSIX_VERSION,
 			      "sysconf returned unexpected value %d",
