+++ tests/posix/headers/prj.conf
@@ -14,7 +14,6 @@ CONFIG_NETWORKING=y
 CONFIG_NET_SOCKETPAIR=y
 
 # for when CONFIG_POSIX_API is not selected
-CONFIG_PTHREAD_IPC=y
 CONFIG_POSIX_FS=y
 CONFIG_POSIX_CLOCK=y
 CONFIG_POSIX_MQUEUE=y
